+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Releases
 
+## 0.10.1
+### Added
+* Improved SpMV performance
+
+## 0.10.0
+### Added
+* CompiledFSM object whichstore the fsm and it's reversal.
+
+## 0.9.1
+### Fixed
+* Do not allocate extra array while doing the backward step.
+
 ## 0.9.0
 ### Added
 * All the different storage of FSM are now replaced with a single type: